Initial commit for mcplotdiff-matplotlib along the lines of html and pyqtgraph variants
