+++ tools/Python/mcplot/matplotlib/CMakeLists.txt
@@ -87,14 +87,33 @@ install(
   RENAME "${P}plot.py"
   )
 
+# Configure fallback script for the a-b difference plotter
+configure_file("${CMAKE_CURRENT_SOURCE_DIR}/mcplotdiff.in" "${WORK}/${P}plotdiff-matplotlib" @ONLY)
+
+# Configure doc for the a-b difference plotter
+configure_file("${CMAKE_CURRENT_SOURCE_DIR}/mcplotdiff-matplotlib.md.in" "${WORK}/${P}plotdiff-matplotlib.md" @ONLY)
+
+# Difference-plotter script, including rename mc/mxplotdiff
+install(
+  PROGRAMS "${PROJECT_SOURCE_DIR}/mcplotdiff.py"
+  DESTINATION "${DEST_TOOLDIR}/${TOOLS_NAME}"
+  RENAME "${P}plotdiff.py"
+  )
+
 # Other .py infrastructure
-# None
+# plotfuncs.py holds the shared, non-CLI plotting/interaction code used by
+# both mcplot.py and mcplotdiff.py; it is NOT renamed (unlike mcplot.py),
+# so both scripts can import it by a fixed name regardless of flavor.
+install(
+  FILES "${PROJECT_SOURCE_DIR}/plotfuncs.py"
+  DESTINATION "${DEST_TOOLDIR}/${TOOLS_NAME}"
+  )
 
 # Wrapper-scripts on windows are handled via
 # cmake Modules/MCUtil
 if(NOT WINDOWS)
   add_custom_target(
-    "CREATE_${PROJECT_NAME}_SYMLINK" ALL DEPENDS "${WORK}/${P}plot-matplotlib"
+    "CREATE_${PROJECT_NAME}_SYMLINK" ALL DEPENDS "${WORK}/${P}plot-matplotlib" "${WORK}/${P}plotdiff-matplotlib"
     )
 
   install(
@@ -110,6 +129,22 @@ if(NOT WINDOWS)
     
   install(
     FILES "${WORK}/${P}plot-matplotlib.md"
+    DESTINATION "${DEST_DATADIR_DOC}"
+  )
+
+  install(
+    PROGRAMS ${PROJECT_SOURCE_DIR}/mcplotdiff.py
+    DESTINATION "${DEST_TOOLDIR}/${TOOLS_NAME}"
+    RENAME "${P}plotdiff.py"
+  )
+
+  install(
+    PROGRAMS "${WORK}/${P}plotdiff-matplotlib"
+    DESTINATION ${DEST_BINDIR}
+    )
+
+  install(
+    FILES "${WORK}/${P}plotdiff-matplotlib.md"
     DESTINATION "${DEST_DATADIR_DOC}"
   )
 endif()
